Add CalculatorHelper for sharing logic
This module includes helper methods that can be shared among different
calculators.
