+++ lib/super_good/solidus_taxjar.rb
@@ -5,6 +5,7 @@ require 'taxjar'
 require "super_good/solidus_taxjar/version"
 require "super_good/solidus_taxjar/api_params"
 require "super_good/solidus_taxjar/api"
+require "super_good/solidus_taxjar/calculator_helper"
 require "super_good/solidus_taxjar/tax_calculator"
 require "super_good/solidus_taxjar/discount_calculator"
